Part6.a: Redux Anecdotes exercises 6.6-6.8
Details:
---
* 6.6 - extract action-creator functions to the reducer (in reality done
  before, but ticks the boxes of this exercise)
* 6.7 - Separate the form for creating new anecdotes into its own
  component
* 6.8 - Separate the code responsible for listing the anecdotes into its
  own component, as an extra create a new component for individual
  anecdotes
