+++ part6/redux-anecdotes/src/App.jsx
@@ -1,47 +1,12 @@
-import { useSelector, useDispatch } from 'react-redux'
-import { incrementVote, createAnecdote } from './reducers/anecdoteReducer'
+import AnecdoteForm from './components/AnecdoteForm'
+import AnecdoteList from './components/AnecdoteList'
 
-const App = () => {
-    const anecdotes = useSelector((state) => state)
-    const dispatch = useDispatch()
-
-    const vote = (id) => {
-        console.log('vote', id)
-        dispatch(incrementVote(id))
-    }
-
-    const addAnecdote = (eve) => {
-        eve.preventDefault()
-        const content = eve.target.newAnecdote.value
-        eve.target.newAnecdote.value = ''
-        dispatch(createAnecdote(content))
-    }
-
-    return (
-        <div>
-            <h2>Anecdotes</h2>
-            {anecdotes
-                .sort((a, b) => b.votes - a.votes)
-                .map((anecdote) => (
-                    <div key={anecdote.id}>
-                        <div>{anecdote.content}</div>
-                        <div>
-                            has {anecdote.votes}
-                            <button onClick={() => vote(anecdote.id)}>
-                                vote
-                            </button>
-                        </div>
-                    </div>
-                ))}
-            <h2>create new</h2>
-            <form onSubmit={addAnecdote}>
-                <div>
-                    <input name="newAnecdote" />
-                </div>
-                <button>create</button>
-            </form>
-        </div>
-    )
-}
+const App = () => (
+    <div>
+        <h2>Anecdotes</h2>
+        <AnecdoteForm />
+        <AnecdoteList />
+    </div>
+)
 
 export default App
