+++ part6/redux-anecdotes/src/components/Anecdote.jsx
@@ -1,0 +1,13 @@
+const Anecdote = ({ anecdote, handleVote }) => {
+    return (
+        <div key={anecdote.id}>
+            <div>{anecdote.content}</div>
+            <div>
+                has {anecdote.votes}
+                <button onClick={handleVote}>vote</button>
+            </div>
+        </div>
+    )
+}
+
+export default Anecdote
